FINAL SUBMISSION on AUT.400
Final version of the submission for a course on informational technology in automation. Done in conjuction with my good peers, to get an analytical look at what it takes to construct a independently functioning, barebones automation process. This process surveys, protects and maintains all of the aspects of IT development in a factory setting, and even has demos in place to help gauge out issues popping up upon deployment. Our group got a top marks from this report, and I'm incredibly proud of our systems.